requirements.txt
flask==3.0.0
paho-mqtt==1.6.1
pyyaml==6.0.1
gunicorn==21.2.0
